+++ lang/en/theme_foundation.php
@@ -53,6 +53,9 @@ $string['prescssdesc'] = 'State the SCSS that should be used before any other SC
 $string['customscss'] = 'Custom SCSS';
 $string['customscssdesc'] = 'Add custom SCSS to the theme.';
 
+$string['hvpcustomcss'] = 'H5P Custom CSS';
+$string['hvpcustomcssdesc'] = 'Custom CSS for the H5P module (moodle.org/plugins/mod_hvp) if installed.';
+
 $string['moduleheading'] = 'Module';
 $string['moduleheadingsub'] = 'Module settings';
 $string['moduleheadingdesc'] = 'Configure the module settings for Foundation here.';